Support multiple variants of the SPDY protocol.
This behavior-free refactoring makes the first baby steps towards
supporting HTTP/2.0. It adds indirection on the framing layer so
we can frame either using SPDY/3's syntax or HTTP/2.0's.
